pyproject.toml
[project]
name = "airflow-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow>=3.1.7",
]
